Integrating the docker module to the archive
The module is far from finished, it just checks
if docker-compose directories are added in the right directory,
and provide a way to add new ones, or remove the existing
ones.

I'm still trying to think on how I could provide a way
to enable these dockers on startup.
Technically, you'd just edit the docker-compose.yml, put
the "restart: always" instruction, up the whole service
and just ensure that docker is available on startup, which
will then remember to start the right containers registered
through docker-compose.

For that workflow, we need to add at least "edit" and "start/stop"
actions.

Another way to handle this would be to register a list of
docker-compose to start on startup, and then have a service
ensure that these dockers are started.
This require additional files that would be very specific to
armbian however, and adding distribution specificities to a
generic technology generally lead to confusion, and hard to
pinpoint issues.

So I'll try to add the "edit" and "start/stop" actions in
the next update.
Then I'd start to focus on the modular design, since that's
still one of non-resolved main point.

Signed-off-by: Myy Miouyouyou <myy@miouyouyou.fr>