Starting to see the beginning of an architecture
The first concept is as follows :

* Inside 'modules' lies every module
  * Each 'module' needs to have at least three files
   * 'module.cli' the CLI version of the module
   * 'DESC' the quick description of the module
   * 'armbian/cli/DEPS' the list of Debian dependencies required
     to execute the software
  * Then for GUI, we'll add 'module.gui' and the appropriate
    armbian version with it.

The idea is that, sometimes, mixing GUI and command lines
do not make sense. There are ways to present information in
a GUI, that makes sense for the software itself,
but cannot translate well into command lines, and
vice-versa.

So the whole point (still not completely done) is :
* When invoking a module through the configurator :
  * If there's no arguments to the module, then
    the configurator will check if it can invoke the
    GUI and do it, if possible.
  * Else, if there's command line arguments
    then the CLI version will be selected by default.

A command line argument or an environment variable should
be able to change that default behaviour, in order to
always execute CLI versions or GUI versions.

Signed-off-by: Myy Miouyouyou <myy@miouyouyou.fr>