docker: Implement a decent UI
It's still not tested. However, the foundations are here.
The whole thing use a UI file that can be edited with
QT Designer easily.

I wanted to use QML but QML comes with too much quirks
and weird dependencies IMHO.
QT5 and PyQT5 dependencies seems much more easier to
install, IMHO.

The DEPS files are not updated at the moment, though.

Signed-off-by: Myy Miouyouyou <myy@miouyouyou.fr>