modules: docker: Small improvements on the UI
Added the ability to check the logs of running instances.

I'm still worried about running background processes with UI,
like "checking the logs continuously" or
"checking the status continuously", since it's the kind of
process that can hog a machine quickly, if they're not handled
correctly when they're not needed.

So, at the moment, I'll just add a few more buttons in order
to check things manually.

Also, docker-compose provide no clear way of checking the
status of an installation. You can use 'ps', and you'll know
the status of each invididual service.
However, it's then up to you to determine if the status of
each service makes sense.

Signed-off-by: Myy Miouyouyou <myy@miouyouyou.fr>