Rework of the folder architecture
In order to package the modules correctly, a package-manager specific
directory architecture has been put in place.

The whole idea to have something like that, inside the module
directory :

* armbian
  * DEPS : Global installation dependencies
  * DEPS.build : Dependencies only required to build the actual
    module project
  * BUILD.sh : The script run at the module root directory to
    build the whole module.
  * POSTINST.sh : Global Post-installation script
  * all-archs/
   * cli/
    * DEPS : Cli specific dependencies
    * module/ : Module files copied on the harddrive...
   * gui/
    * POSTINST.sh : GUI specific Post-installation script
    * DEPS : GUI specific dependencies
    * module/ : Module files copied on the harddrive...
  * armv7/ : ARMv7 specific modules
  * aarch64 : ARMv8 64 bits specific modules
  * aarch64.rockchip : ARMv8 64 bits specific modules for Rockchip

First, note that the architectures directories are not required
per-se. You can go with only "all-archs". Only create an "armv7"
directory if it makes sense for you to have a Armv7 specific
module.
This is generally the case if the module needs to be built for
a specific architecture.

Second, while the files (DEPS, POSTINST.sh, ...) are concatenated,
each architecture is independant, and the packaging script WILL NOT
OVERLAY THEM.
Meaning that the module files need to copied over and over, in each
invidiual architecture directory.

Signed-off-by: Myy Miouyouyou <myy@miouyouyou.fr>